% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals}
\alias{animals}
\title{Leche materna de animales}
\format{
Data con 25 registros y 5 variables
\describe{
\item{water}{Porcentaje de agua que compone la leche materna.}
\item{protein}{Porcentaje de agua que compone la leche materna.}
\item{fat}{Porcentaje de grasa que compone la leche materna.}
\item{lactose}{Porcentaje de lactosa que compone la leche materna.}
\item{ash}{Porcentaje de ceniza que compone la leche materna.}
}
}
\usage{
  animals
}
\description{
  Esta data contiene información de cinco componentes de la leche materna de 25 especies de animales.
}
\examples{
  data(animals)
}
\keyword{datasets}