% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apartamentos}
\alias{apartamentos}
\title{Precios de apartamentos}
\format{
Data con 141 registros y 2 variables
\describe{
\item{precio}{Precio del apartamento en pesos colombianos}
\item{area.cons}{Area construida del apartamento en metros cuadrados}
}
}
\usage{
apartamentos
}
\description{
Esta data contiene el precio y area construida de apartamentos.
}
\examples{
data(apartamentos)
}
\keyword{datasets}