% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beer2}
\alias{beer2}
\title{Caracteristicas de cervezas}
\format{
Data con 69 registros y 3 variables
\describe{
\item{precio}{Precio de la cerveza en dólares}
\item{calorias}{Calorias de la cerveza}
\item{poralcoh}{Porcentaje de alcohol}
}
}
\usage{
  beer2
}
\description{
  Esta data contiene caracteristicas 69 tipos de cerveza.
}
\examples{
  data(beer2)
}
\keyword{datasets}
