% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cafe}
\alias{cafe}
\title{Puntajes de cafés (ACP)}
\format{
Data con 10 registros y 7 variables
\describe{
\item{Iden}{Identificador único del café.}
\item{IAroma}{Puntaje del aroma antes de revolver el café.}
\item{Aroma}{Puntaje del aroma despues de revolver el café.}
\item{Cuerpo}{Puntaje del cuerpo.}
\item{Acidez}{Puntaje de la acidez.}
\item{Amargo}{Puntaje de la amargura.}
\item{Astring}{Puntaje de la astringencia.}
}
}
\usage{
  cafe
}
\description{
  Esta data contiene información sobre distintos tipos de café. Todos los puntajes están en un rango de 1 a 10.
}
\examples{
  data(cafe)
}
\keyword{datasets}
