% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cafe01}
\alias{cafe01}
\title{Puntajes de cafés (ACP)}
\format{
Data con 10 registros y 12 variables
\describe{
\item{id}{Identificador único del café.}
\item{especie}{Especie.}
\item{pais}{País de proveniencia.}
\item{aroma}{Puntaje del aroma.}
\item{sabor}{Puntaje del sabor.}
\item{regusto}{Puntaje general despues de probar el café por segunda vez.}
\item{acidez}{Puntaje de la acidez.}
\item{cuerpo}{Puntaje del cuerpo.}
\item{balance}{Puntaje del balance.}
\item{dulzura}{Puntaje de la dulzura.}
\item{puntoscop}{Puntaje cop del café.}
\item{totalpunt}{Suma total de los puntajes.}
}
}
\usage{
  cafe01
}
\description{
  Esta data contiene información sobre distintos tipos de café. Todos los puntajes están en un rango de 1 a 10.
}
\examples{
  data(cafe01)
}
\keyword{datasets}
