% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casas}
\alias{casas}
\title{Caracteristicas de casas.}
\format{
Data con 54 registros y 10 variables
\describe{
\item{id}{Identificador único de la casa.}
\item{precio}{Precio en millones de pesos.}
\item{imp.pred}{Impuesto predial anual en miles de pesos.}
\item{num.ban}{Número de baños.}
\item{area.lot}{Área total del lote en metros cuadrados.}
\item{area.con}{Área construida en metros cuadrados.}
\item{num.gar}{Número de parqueaderos.}
\item{num.cua}{Número de ventanas.}
\item{num.hab}{Número de habitaciones.}
\item{edad}{Edad en años.}
}
}
\usage{
  casas
}
\description{
  Esta data extraida mediante web scraping contiene caracteristicas de casas.
}
\examples{
  data(casas)
}
\keyword{datasets}