% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataMat}
\alias{dataMat}
\title{Notas del último parcial de mátematcas}
\format{
Data con 200 registros y 2 variables
\describe{
\item{gana}{Indica si el estudiante pasó la materia de mátematicas}
\item{nota}{Nota del último parcial de mátematicas}
}
}
\usage{
  dataMat
}
\description{
  Esta data contiene las notas del último parcial de mátematicas de estudiantes de universidad.
}
\examples{
  data(dataMat)
}
\keyword{datasets}