% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Empleados de compañía}
\format{
Data con 1470 registros y 5 variables
\describe{
\item{Rotación}{Indica si el trabajador ha sido cambiado de departamento.}
\item{Edad}{Edad del empleado.}
\item{Viaje de negocios}{Indica con que frecuencia el empleado viaja debido a negocios.}
\item{Departamento}{Departamento a la que pertenece el empleado.}
\item{Distancia_Casa}{Distancia en kilometros de la casa del empleado hasta la oficina.}
}
}
\usage{
  datos
}
\description{
  Esta data contiene información sobre trabajadores de una compañía.
}
\examples{
  data(datos)
}
\keyword{datasets}