% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflacion2022}
\alias{inflacion2022}
\title{Inflación en Colombia hasta el año 2022.}
\format{
Data con 141 registros y 2 variables
\describe{
\item{id}{Identificador único del registro.}
\item{fecha}{Fecha.}
\item{inflacion}{Porcentaje de inflación con respecto al mes del año anterior.}
\item{anio}{Año.}
\item{dia}{Variable nula, en este caso no se usara.}
\item{mes}{Mes.}
}
}
\usage{
  inflacion2022
}
\description{
  Esta data contiene la inflación por mes y año en Colombia hasta el 2022.
}
\examples{
  data(inflacion2022)
}
\keyword{datasets}
