% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventarioPackages}
\alias{inventarioPackages}
\title{Tópicos estadísticos}
\format{
Data con 19342 registros y 3 variables
\describe{
\item{date}{Fecha de publicación}
\item{Package}{Nombre del paquete que contiene el tópico}
\item{Title}{Tópico, este puede ser un artículo, método estadístico o base de datos.}
}
}
\usage{
  inventarioPackages
}
\description{
  Esta base de datos contiene diversos tópicos y métodos estadísticos que se pueden encontrar en paquetes de R.
}
\examples{
  data(inventarioPackages)
}
\keyword{datasets}