% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matricula}
\alias{matricula}
\title{Notas de matem치ticas}
\format{
Data con 200 registros y 2 variables
\describe{
\item{honor}{Indica si el estudiante se gradu칩 con honores}
\item{matematicas}{Puntaje obtenido en matem치ticas}
}
}
\usage{
  matricula
}
\description{
  Esta data contiene informaci칩n sobre estudiantes graduados de un colegio.
}
\examples{
  data(matricula)
}
\keyword{datasets}