% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socialnet}
\alias{socialnet}
\title{Compras en red social}
\format{
Data con 400 registros y 5 variables
\describe{
\item{User ID}{Identificador único del usuario}
\item{Gender}{Género.}
\item{Age}{Edad en años.}
\item{EstimatedSalary}{Salario anual en dólares estimado.}
\item{Purchased}{Indica si la persona compró un producto, un 1 indica que si realizó al menos una compra.}
}
}
\usage{
  socialnet
}
\description{
  Esta data contiene información sobre personas que han comprado productos en una red social.
}
\examples{
  data(socialnet)
}
\keyword{datasets}