% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ventas}
\alias{ventas}
\title{Ventas de empleados}
\format{
Data con 15 registros y 3 variables
\describe{
\item{Vendedor}{Identificador único del vendedor}
\item{Clientes}{Número de clientes del vendedor}
\item{Ventas}{Número de de ventas del vendedor}
}
}
\usage{
  ventas
}
\description{
  Esta data contiene información sobre las ventas de empleados de una compañía.
}
\examples{
  data(ventas)
}
\keyword{datasets}