% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vivienda1}
\alias{vivienda1}
\title{Viviendas de Cali}
\format{
Data con 8322 registros y 5 variables
\describe{
\item{zona}{Zona de la ciudad en la que está ubicada la vivienda.}
\item{estrato}{Estrato.}
\item{preciom}{Precio en millones de pesos.}
\item{areaconst}{Área construidaen metros cuadrados.}
\item{tipo}{Tipo de vivienda.}
}
}
\usage{
  vivienda1
}
\description{
  Esta data contiene información sobre caracteristicas de viviendas de la ciudad de Cali.
}
\examples{
  data(vivienda1)
}
\keyword{datasets}