% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whisky}
\alias{whisky}
\title{Tipos de whisky (ACP y clustering)}
\format{
Data con 35 registros y 5 variables
\describe{
\item{IND}{Identificador único del whisky.}
\item{PRECIO}{Precio en dólares.}
\item{PROP}{Proporción de alcohol.}
\item{AÑEJO}{Años de añejo.}
\item{NOTA}{Nota general del whisky (0-5)}
}
}
\usage{
  whisky
}
\description{
  Esta data contiene información sobre distintos tipos de whisky.
}
\examples{
  data(whisky)
}
\keyword{datasets}
